Eliminate import cycles (IDRopRowInfo and IDocumentAddTileOptions)
