+++ src/models/document/document-types.ts
@@ -59,9 +59,3 @@ export interface IDocumentContext {
   getProperty: (key: string) => string | undefined;
   setProperties: (properties: ISetProperties) => void;
 }
-
-export interface IDocumentAddTileOptions {
-  title?: string;
-  addSidecarNotes?: boolean;
-  url?: string;
-}
